requirements.txt
flask>=2.3.0
tensorflow>=2.13.0
Pillow>=10.0.0
numpy>=1.24.0
werkzeug>=2.3.0
pytest>=7.4.0
